+++ spec/lib/hyperion/formats_spec.rb
@@ -1,4 +1,7 @@
 require 'spec_helper'
+require 'time'
+
+
 
 class Hyperion
   describe Formats do
@@ -21,8 +24,22 @@ class Hyperion
       end
 
       context 'formats' do
-        it 'writes json' do
-          expect(write({'a' => 1}, :json)).to be_json_eql '{"a":1}'
+        context 'when writing json' do
+          it 'writes hashes with string keys' do
+            expect(write({'a' => 1}, :json)).to be_json_eql '{"a":1}'
+          end
+          it 'writes hashes with symbol keys' do
+            expect(write({a: 1}, :json)).to be_json_eql '{"a":1}'
+          end
+          context 'when writing times' do
+            let!(:time) { Time.parse('2015-02-13 08:40:20.321 -0500') }
+            it 'writes Time objects in UTC ISO 8601 format with milliseconds precision' do
+              expect(write({'a' => time}, :json)).to be_json_eql '{"a":"2015-02-13T13:40:20.321Z"}'
+            end
+            it 'preserves default behavior for non-hyperion code' do
+              expect(Oj.dump({'a' => time})).to be_json_eql '{"a":{"^t":1423834820.321}}'
+            end
+          end
         end
         it 'allows protobuf format but just passes it through' do
           expect(write('x', :protobuf)).to eql 'x'
@@ -37,8 +54,15 @@ class Hyperion
       it 'returns the input if format is nil' do
         expect(read('abc', nil)).to eql 'abc'
       end
-      it 'accepts format as a symbol' do
-        expect(read('{"a":1}', :json)).to eql({'a' => 1})
+      context 'when reading json' do
+        it 'accepts format as a symbol' do
+          expect(read('{"a":1}', :json)).to eql({'a' => 1})
+        end
+        it 'reads times as strings' do
+          # sadly, Oj does not appear to support reading real Time objects, even when using
+          # its own representation ({"^t":1423834820.321})
+          expect(read('{"a":"2015-02-13T13:40:20.321Z"}', :json)['a']).to eql '2015-02-13T13:40:20.321Z'
+        end
       end
       it 'accepts format as a descriptor' do
         descriptor = double(format: :json)